split wizards in separate files
